+++ mobile/cmp/navigator/NavigatorPageModel.js
@@ -4,9 +4,12 @@
  *
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
-import {HoistModel} from '@xh/hoist/core';
+import {HoistModel, managed} from '@xh/hoist/core';
+import {computed} from '@xh/hoist/mobx';
 import {withDefault} from '@xh/hoist/utils/js';
 import {stringify} from 'qs';
+
+import {NavigatorRefreshContextModel} from './impl/NavigatorRefreshContextModel';
 
 /**
  * Model for a NavigatorPage within a Navigator. Specifies the actual content (i.e. page)
@@ -24,6 +27,9 @@ export class NavigatorPageModel {
     disableDirectLink;
     disableAppRefreshButton;
 
+    navigatorModel;
+    @managed refreshContextModel;
+
     /**
      * A generated key which combines the id and sorted props to identify the page in the stack.
      * The returned string is stable for a given id and props.
@@ -34,26 +40,55 @@ export class NavigatorPageModel {
         return stringify({id, props}, qsOpts);
     }
 
+    get renderMode() {
+        return this._renderMode || this.navigatorModel.renderMode;
+    }
+
+    get refreshMode() {
+        return this._refreshMode || this.navigatorModel.refreshMode;
+    }
+
+    @computed
+    get isActive() {
+        const current = this.navigatorModel.getCurrentPageModel();
+        return current?.id === this.id;
+    }
+
     /**
      * @param {string} id - unique ID. Must match a configured Router5 route name.
-     * @param {(Object|function)} c.content - Hoist Component (class or functional) to be
+     * @param {NavigatorModel} navigatorModel - parent NavigatorModel. Provided by the
+     *      navigator when constructing these models - no need to specify manually.
+     * @param {(Object|function)} content - Hoist Component (class or functional) to be
      *      rendered by this page; or function returning react element to be rendered by this page.
      * @param {Object} [props] - props to be passed to page upon creation.
+     * @param {RefreshMode} [renderMode] - strategy for rendering this Page. If null, will
+     *      default to its Navigator's mode. See enum for description of supported modes.
+     * @param {RefreshMode} [refreshMode] - strategy for refreshing this Page. If null, will
+     *      default to its Navigator's mode. See enum for description of supported modes.
      * @param {boolean} [disableDirectLink] - Don't allow the route can be arrived at in a new browser session.
      *      Non-linkable routes are unwound to a safe starting point at the start of a new session.
      * @param {boolean} [disableAppRefreshButton] - Hide any visible app refresh button when at this page.
      */
     constructor({
         id,
+        navigatorModel,
         content,
         props,
+        renderMode,
+        refreshMode,
         disableDirectLink,
         disableAppRefreshButton
     }) {
         this.id = id;
+        this.navigatorModel = navigatorModel;
         this.content = content;
         this.props = withDefault(props, {});
         this.disableDirectLink = withDefault(disableDirectLink, false);
         this.disableAppRefreshButton = withDefault(disableAppRefreshButton, false);
+
+        this._renderMode = renderMode;
+        this._refreshMode = refreshMode;
+
+        this.refreshContextModel = new NavigatorRefreshContextModel(this);
     }
 }